fix: Implement workaround for racey `listsinceblock`
It seems that `listsinceblock` is not atomic and could provide results
that are inconsistent with each other. The workaround checks if the
chain tip moved while reading `listsinceblock` and retries if it did.

This is similar to how atomicity is achieved for `listunspent` (in
`Query::list_unspent_raw`) and previously for `listtransactions`
(before being replaced by `listsinceblock`).

The bug was reported here: https://github.com/bitcoin/bitcoin/issues/19338