Improve build times with the `http` feature by 65% (!)
From around 6 minutes to 2 minutes.

Accomplished by building a more balanced `Or` tree of warp filters.

The tree being built is not optimally balanced, because the code also
optimizes for readability. This could be further improved with a macro
that creates the optimally balanced tree, which would be more easily
maintainable and readable compared to hand-crafting it.

Refs: #15, https://github.com/seanmonstar/warp/issues/619