Build nodejs-bwt-daemon after building all platforms
Prepare the node package in the higher-level release.sh script, after
building libbwt for all the platforms via build.sh. Needed because the
node package includes the SHA256SUMS of the libbwt distribution files.

This commit also removes nodejs-bwt-daemon from the reproducible CI
build process in Travis, because including the hashes for all platforms
is not possible with the current Travis setup, which builds each
platform separately (and concurrently).

This should eventually be made to work on Travis, but better not to
include the node package at all for now to avoid potential confusion.