ffi: Allow shutting down the daemon while starting up
The shutdown handler is now passed to a new `init_cb` callback,
which can be used to shutdown the daemon without waiting for it
to finish starting up first.

If the shutdown signal is received while importing/rescanning addresses,
the daemon will not stop immediately but will be marked for later termination.